Refactor: move processPCB to own file
